requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
dlib==19.19.0
docutils==0.16
face-recognition==1.3.0
face-recognition-models==0.3.0
idna==2.9
Kivy==1.11.1
Kivy-Garden==0.1.4
numpy==1.18.1
opencv-python==4.2.0.32
Pillow==7.0.0
Pygments==2.5.2
requests==2.23.0
urllib3==1.25.8
